Get rid of internal transform operator
