+++ reaktive/src/commonMain/kotlin/com/badoo/reaktive/base/ErrorCallbackExt.kt
@@ -1,0 +1,11 @@
+package com.badoo.reaktive.base
+
+internal inline fun <T> ErrorCallback.tryCatch(block: () -> T, onSuccess: (T) -> Unit) {
+    try {
+        block()
+    } catch (e: Throwable) {
+        onError(e)
+        return
+    }
+        .also(onSuccess)
+}
