+++ reaktive/src/commonMain/kotlin/com/badoo/reaktive/maybe/Filter.kt
@@ -1,10 +1,23 @@
 package com.badoo.reaktive.maybe
 
+import com.badoo.reaktive.base.Observer
+import com.badoo.reaktive.base.subscribeSafe
+import com.badoo.reaktive.base.tryCatch
+import com.badoo.reaktive.completable.CompletableCallbacks
+
 fun <T> Maybe<T>.filter(predicate: (T) -> Boolean): Maybe<T> =
-    transform { value, onSuccess, onComplete ->
-        if (predicate(value)) {
-            onSuccess(value)
-        } else {
-            onComplete()
-        }
+    maybeUnsafe { observer ->
+        subscribeSafe(
+            object : MaybeObserver<T>, Observer by observer, CompletableCallbacks by observer {
+                override fun onSuccess(value: T) {
+                    observer.tryCatch({ predicate(value) }) {
+                        if (it) {
+                            observer.onSuccess(value)
+                        } else {
+                            observer.onComplete()
+                        }
+                    }
+                }
+            }
+        )
     }
