+++ reaktive/src/commonMain/kotlin/com/badoo/reaktive/maybe/Map.kt
@@ -1,6 +1,17 @@
 package com.badoo.reaktive.maybe
 
+import com.badoo.reaktive.base.Observer
+import com.badoo.reaktive.base.subscribeSafe
+import com.badoo.reaktive.base.tryCatch
+import com.badoo.reaktive.completable.CompletableCallbacks
+
 fun <T, R> Maybe<T>.map(mapper: (T) -> R): Maybe<R> =
-    transform { value, onSuccess, _ ->
-        onSuccess(mapper(value))
+    maybeUnsafe { observer ->
+        subscribeSafe(
+            object : MaybeObserver<T>, Observer by observer, CompletableCallbacks by observer {
+                override fun onSuccess(value: T) {
+                    observer.tryCatch({ mapper(value) }, observer::onSuccess)
+                }
+            }
+        )
     }
