+++ reaktive/src/commonMain/kotlin/com/badoo/reaktive/maybe/NotNull.kt
@@ -1,10 +1,20 @@
 package com.badoo.reaktive.maybe
 
+import com.badoo.reaktive.base.Observer
+import com.badoo.reaktive.base.subscribeSafe
+import com.badoo.reaktive.completable.CompletableCallbacks
+
 fun <T : Any> Maybe<T?>.notNull(): Maybe<T> =
-    transform { value, onSuccess, onComplete ->
-        if (value != null) {
-            onSuccess(value)
-        } else {
-            onComplete()
-        }
+    maybeUnsafe { observer ->
+        subscribeSafe(
+            object : MaybeObserver<T?>, Observer by observer, CompletableCallbacks by observer {
+                override fun onSuccess(value: T?) {
+                    if (value != null) {
+                        observer.onSuccess(value)
+                    } else {
+                        observer.onComplete()
+                    }
+                }
+            }
+        )
     }
