+++ reaktive/src/commonMain/kotlin/com/badoo/reaktive/observable/Map.kt
@@ -1,6 +1,21 @@
 package com.badoo.reaktive.observable
 
+import com.badoo.reaktive.base.subscribeSafe
+import com.badoo.reaktive.base.tryCatch
+import com.badoo.reaktive.completable.CompletableCallbacks
+import com.badoo.reaktive.disposable.Disposable
+
 fun <T, R> Observable<T>.map(mapper: (T) -> R): Observable<R> =
-    transform { value, onNext ->
-        onNext(mapper(value))
+    observable { emitter ->
+        subscribeSafe(
+            object : ObservableObserver<T>, CompletableCallbacks by emitter {
+                override fun onSubscribe(disposable: Disposable) {
+                    emitter.setDisposable(disposable)
+                }
+
+                override fun onNext(value: T) {
+                    emitter.tryCatch({ mapper(value) }, emitter::onNext)
+                }
+            }
+        )
     }
